Interface bases separated
