+++ interface/src/App/Pokedex/PokeCard/SliderBody/EffBlock/EffIcon.jsx
@@ -6,7 +6,7 @@ import Box from '@material-ui/core/Box';
 import PropTypes from 'prop-types';
 
 import Iconer from "App/Components/Iconer/Iconer";
-import { typeDecoder } from "js/indexFunctions";
+import { typeDecoder } from "js/coders/typeDecoder";
 
 const EffIcon = React.memo(function EffIcon(props) {
     return (
