+++ interface/src/js/indexFunctions.js
@@ -1,6 +1,9 @@
 import { levelData } from "js/bases/levelData";
 import { tierHP } from "js/bases/tierHP";
 import { weather } from "js/bases/weather";
+import { friendship } from "js/bases/friendship";
+import { effectivenessData } from "js/bases/effectivenessData";
+import { stagesData } from "js/bases/stagesData";
 
 export const capitalizeFirst = (str, lower = false) =>
     (lower ? str.toLowerCase() : str).replace(/(?:^|\s|-|["'([{])+\S/g, match => match.toUpperCase());
@@ -689,440 +692,9 @@ export function returnVunStyle(rate) {
     return "4"
 }
 
-export var typeDecoder = [
-    "Bug",
-    "Dark",
-    "Dragon",
-    "Electric",
-    "Fairy",
-    "Fighting",
-    "Fire",
-    "Flying",
-    "Ghost",
-    "Grass",
-    "Ground",
-    "Ice",
-    "Normal",
-    "Poison",
-    "Psychic",
-    "Rock",
-    "Steel",
-    "Water",
-]
 
-export var typeEncoder = {
-    "Bug": 0,
-    "Dark": 1,
-    "Dragon": 2,
-    "Electric": 3,
-    "Fairy": 4,
-    "Fighting": 5,
-    "Fire": 6,
-    "Flying": 7,
-    "Ghost": 8,
-    "Grass": 9,
-    "Ground": 10,
-    "Ice": 11,
-    "Normal": 12,
-    "Poison": 13,
-    "Psychic": 14,
-    "Rock": 15,
-    "Steel": 16,
-    "Water": 17,
-}
 
-var stagesData = {
-    "-4": 0.5,
-    "-3": 0.57,
-    "-2": 0.67,
-    "-1": 0.8,
-    "0": 1,
-    "1": 1.25,
-    "2": 1.5,
-    "3": 1.75,
-    "4": 2,
-}
 
-export var effectivenessData = [
-    [
-        0,
-        1.6,
-        0,
-        0,
-        0.625,
-        0.625,
-        0.625,
-        0.625,
-        0.625,
-        1.6,
-        0,
-        0,
-        0,
-        0.625,
-        1.6,
-        0,
-        0.625,
-        0
-    ],
-    [
-        0,
-        0.625,
-        0,
-        0,
-        0.625,
-        0.625,
-        0,
-        0,
-        1.6,
-        0,
-        0,
-        0,
-        0,
-        0,
-        1.6,
-        0,
-        0,
-        0
-    ],
-    [
-        0,
-        0,
-        1.6,
-        0,
-        0.391,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0.625,
-        0
-    ],
-    [
-        0,
-        0,
-        0.625,
-        0.625,
-        0,
-        0,
-        0,
-        1.6,
-        0,
-        0.625,
-        0.391,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        1.6
-    ],
-    [
-        0,
-        1.6,
-        1.6,
-        0,
-        0,
-        1.6,
-        0.625,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0.625,
-        0,
-        0,
-        0.625,
-        0
-    ],
-    [
-        0.625,
-        1.6,
-        0,
-        0,
-        0.625,
-        0,
-        0,
-        0.625,
-        0.391,
-        0,
-        0,
-        1.6,
-        1.6,
-        0.625,
-        0.625,
-        1.6,
-        1.6,
-        0
-    ],
-    [
-        1.6,
-        0,
-        0.625,
-        0,
-        0,
-        0,
-        0.625,
-        0,
-        0,
-        1.6,
-        0,
-        1.6,
-        0,
-        0,
-        0,
-        0.625,
-        1.6,
-        0.625
-    ],
-    [
-        1.6,
-        0,
-        0,
-        0.625,
-        0,
-        1.6,
-        0,
-        0,
-        0,
-        1.6,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0.625,
-        0.62,
-        0
-    ],
-    [
-        0,
-        0.625,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        1.6,
-        0,
-        0,
-        0,
-        0.391,
-        0,
-        1.6,
-        0,
-        0,
-        0
-    ],
-    [
-        0.625,
-        0,
-        0.625,
-        0,
-        0,
-        0,
-        0.625,
-        0.625,
-        0,
-        0.625,
-        1.6,
-        0,
-        0,
-        0.625,
-        0,
-        1.6,
-        0.625,
-        1.6
-    ],
-    [
-        0.625,
-        0,
-        0,
-        1.6,
-        0,
-        0,
-        1.6,
-        0.391,
-        0,
-        0.625,
-        0,
-        0,
-        0,
-        1.6,
-        0,
-        1.6,
-        1.6,
-        0
-    ],
-    [
-        0,
-        0,
-        1.6,
-        0,
-        0,
-        0,
-        0.625,
-        1.6,
-        0,
-        1.6,
-        1.6,
-        0.625,
-        0,
-        0,
-        0,
-        0,
-        0.625,
-        0.625
-    ],
-    [
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0.391,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0.625,
-        0.625,
-        0
-    ],
-    [
-        0,
-        0,
-        0,
-        0,
-        1.6,
-        0,
-        0,
-        0,
-        0.625,
-        1.6,
-        0.625,
-        0,
-        0,
-        0.625,
-        0,
-        0.625,
-        0.391,
-        0
-    ],
-    [
-        0,
-        0.391,
-        0,
-        0,
-        0,
-        1.6,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        1.6,
-        0.625,
-        0,
-        0.625,
-        0
-    ],
-    [
-        1.6,
-        0,
-        0,
-        0,
-        0,
-        0.625,
-        1.6,
-        1.6,
-        0,
-        0,
-        0.625,
-        1.6,
-        0,
-        0,
-        0,
-        0,
-        0.625,
-        0
-    ],
-    [
-        0,
-        0,
-        0,
-        0.625,
-        1.6,
-        0,
-        0.625,
-        0,
-        0,
-        0,
-        0,
-        1.6,
-        0,
-        0,
-        0,
-        1.6,
-        0.625,
-        0.625
-    ],
-    [
-        0,
-        0,
-        0.625,
-        0,
-        0,
-        0,
-        1.6,
-        0,
-        0,
-        0.625,
-        1.6,
-        0,
-        0,
-        0,
-        0,
-        1.6,
-        0,
-        0.625
-    ]
-]
 
-export var tierMult = [
-    0.5974,
-    0.67,
-    0.73,
-    0.79,
-    0.79,
-    0.79,
-]
 
-var friendship = [
-    1.0,
-    1.03,
-    1.05,
-    1.07,
-    1.1,
-    1.06,
-    1.12,
-    1.18,
-    1.25,
-]
+
